EIP-7 Changed location of fun get_db to database.py and divided endpoints to better organize the code
